refactor: constants.ts のドメイン別分割（P5）
- constants/ ディレクトリを新設し14ドメイン別定数ファイルに分割
  - battle, evolution, biome, difficulty, skill, event, tree,
    achievement, awakening, ally, ui, save, scaling, audio
- constants/index.ts: barrel export（46定数を統合エクスポート）
- constants.ts を barrel re-export に変換（後方互換性維持）
- TB_KEY_COLOR を TreeScreen.tsx のローカル定義から constants/ui.ts に移動
- TreeScreen.tsx の不要な TreeBonus 型インポートを削除
- 新規テスト: 15テストファイル / 77テスト追加
- 全274スイート / 3552テストパス
