+++ src/features/primal-path/components/TreeScreen.tsx
@@ -1,19 +1,9 @@
 import React from 'react';
-import type { SaveData, SfxType, TreeBonus } from '../types';
+import type { SaveData, SfxType } from '../types';
 import type { GameAction } from '../hooks';
-import { TREE, TIER_UNLOCK, TIER_NAMES, CAT_CL, TB_SUMMARY } from '../constants';
+import { TREE, TIER_UNLOCK, TIER_NAMES, CAT_CL, TB_SUMMARY, TB_KEY_COLOR } from '../constants';
 import { getTB } from '../game-logic';
 import { Screen, SubTitle, Divider, GameButton, TreeNodeBox, TierHeader } from '../styles';
-
-/** ツリーボーナスキーからカテゴリカラーへのマッピング */
-const TB_KEY_COLOR: Readonly<Record<keyof TreeBonus, string>> = {
-  bA: CAT_CL.atk, bH: CAT_CL.hp, bD: CAT_CL.def,
-  cr: CAT_CL.crit, bM: CAT_CL.bone, dM: CAT_CL.atk,
-  rg: CAT_CL.hp, rv: CAT_CL.spc, iR: CAT_CL.env,
-  fR: CAT_CL.env, aS: CAT_CL.ally, aH: CAT_CL.ally,
-  aA: CAT_CL.ally, eN: CAT_CL.spc, sC: CAT_CL.spc,
-  rr: CAT_CL.hp, fQ: CAT_CL.spc, aQ: CAT_CL.ally, rP: CAT_CL.spc,
-};
 
 interface Props {
   save: SaveData;
