+++ src/features/primal-path/constants/index.ts
@@ -1,0 +1,18 @@
+/**
+ * 定数の barrel export
+ * 全ドメイン別定数を統合エクスポート
+ */
+export { ENM, BOSS, BOSS_CHAIN_SCALE, FINAL_BOSS_ORDER, SPEED_OPTS, WAVES_PER_BIOME, ENEMY_COLORS, ENEMY_DETAILS, ENEMY_SMALL_DETAILS } from './battle';
+export { EVOS, SYNERGY_BONUSES, SYNERGY_TAG_INFO } from './evolution';
+export { BIO, BIOME_COUNT, BIOME_AFFINITY, ENV_DMG } from './biome';
+export { DIFFS } from './difficulty';
+export { A_SKILLS, SFX_DEFS } from './skill';
+export { RANDOM_EVENTS, EVENT_CHANCE, EVENT_MIN_BATTLES } from './event';
+export { TREE, TIER_UNLOCK, TIER_NAMES } from './tree';
+export { ACHIEVEMENTS, CHALLENGES } from './achievement';
+export { AWK_SA, AWK_FA } from './awakening';
+export { ALT } from './ally';
+export { CIV_TYPES, CIV_KEYS, TC, TN, CAT_CL, LOG_COLORS, TB_SUMMARY, TB_DEFAULTS, TB_KEY_COLOR } from './ui';
+export { FRESH_SAVE, SAVE_KEY, STATS_KEY, ACHIEVEMENTS_KEY, AGGREGATE_KEY, MAX_RUN_STATS } from './save';
+export { LOOP_SCALE_FACTOR, ENDLESS_LINEAR_SCALE, ENDLESS_EXP_BASE, ENDLESS_AM_REFLECT_RATIO } from './scaling';
+export { BGM_PATTERNS, VOLUME_KEY } from './audio';
