+++ src/features/primal-path/constants/scaling.ts
@@ -1,0 +1,15 @@
+/**
+ * スケーリング関連の定数
+ */
+
+/** 周回倍率係数 */
+export const LOOP_SCALE_FACTOR = 0.5;
+
+/** エンドレスモード: 線形スケール係数（wave あたり） */
+export const ENDLESS_LINEAR_SCALE = 0.18;
+
+/** エンドレスモード: 指数スケール底 */
+export const ENDLESS_EXP_BASE = 1.15;
+
+/** エンドレスモード: プレイヤーaM超過分の敵スケーリング反映率 */
+export const ENDLESS_AM_REFLECT_RATIO = 0.5;
